agent/requirements.txt
livekit-agents>=0.12.11,<1.0.0
livekit-plugins-openai>=0.10.17,<1.0.0
livekit-plugins-cartesia>=0.4.7,<1.0.0
livekit-plugins-deepgram>=0.6.17,<1.0.0
livekit-plugins-silero>=0.7.4,<1.0.0
livekit-plugins-turn-detector>=0.4.0,<1.0.0
python-dotenv~=1.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
prisma>=0.9.0
python-socketio[asyncio]>=5.8.0

# Set WEBSOCKET_URL environment variable to match backend-express config
# e.g., WEBSOCKET_URL=http://localhost:5000
